AWMProjects/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-filter==24.3
django-pwa==2.0.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
idna==3.10
map==1.3.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
